requirements.txt
torch~=1.11.0
timm~=0.6.7
natsort~=8.2.0
opencv-python~=4.6.0.66
torchvision~=0.12.0
pillow~=9.1.1
pyyaml~=6.0
tqdm~=4.64.1
numpy~=1.22.1
tensorboardx~=2.5.1
